fix: discount code lost focus after typing
- further split billable metricsData component
